move display register writing code from core back into display
