+++ sensehat.h
@@ -61,6 +61,4 @@ enum gamma_preset {
 	GAMMA_PRESET_COUNT,
 };
 
-int sensehat_update_display(struct sensehat *sensehat);
-
 #endif
